requirements.txt
pyaml
influxdb-client